Combined related data classes into single files, added missing obfuscation annotations.
