--- shared/src/commonMain/kotlin/com/feko/generictabletoprpg/shared/features/tracker/model/EquipmentContainer.kt
@@ -1,14 +1,0 @@
-package com.feko.generictabletoprpg.shared.features.tracker.model
-
-import com.feko.generictabletoprpg.shared.common.domain.model.DoNotObfuscate
-import kotlinx.serialization.Serializable
-
-@Serializable
-@DoNotObfuscate
-data class EquipmentContainer(
-    val entries: List<EquipmentEntry>
-) {
-    companion object {
-        val Empty = EquipmentContainer(listOf())
-    }
-}
