--- shared/src/commonMain/kotlin/com/feko/generictabletoprpg/shared/features/tracker/model/IEquipmentItem.kt
@@ -1,5 +1,0 @@
-package com.feko.generictabletoprpg.shared.features.tracker.model
-
-import com.feko.generictabletoprpg.shared.common.domain.model.INamed
-
-interface IEquipmentItem : INamed
