--- shared/src/commonMain/kotlin/com/feko/generictabletoprpg/shared/features/tracker/model/StatSkillEntry.kt
@@ -1,16 +1,0 @@
-package com.feko.generictabletoprpg.shared.features.tracker.model
-
-import com.feko.generictabletoprpg.shared.common.domain.model.DoNotObfuscate
-import kotlinx.serialization.Serializable
-
-@DoNotObfuscate
-@Serializable
-data class StatSkillEntry(
-    val name: String,
-    val bonus: Int,
-    val passiveScore: Int,
-    val additionalBonus: Int,
-    val isProficient: Boolean,
-    val hasExpertise: Boolean = false,
-    val showPassive: Boolean = false
-)
